// src/generated/api/trading/schema.ts
export interface components {
    schemas: {
        CreateOrderRequest: {
            idempotency_key: string;
            quote_id: string;
        };
        OrderDetail: {
            approval_reason?: string;
            /** Format: date-time */
            approved_at?: string;
            approved_by?: string;
            /** Format: int64 */
            base_amount: number;
            base_asset_code: string;
            base_unit_code: string;
            /** Format: date-time */
            cancelled_at?: string;
            /** Format: date-time */
            completed_at?: string;
            /** Format: date-time */
            created_at: string;
            execution_id?: string;
            failure_reason?: string;
            id: string;
            idempotency_key: string;
            market_symbol: string;
            price_snapshot_id: string;
            quote_asset_code: string;
            quote_id: string;
            side: string;
            status: string;
            /** Format: int64 */
            total_quote_amount: number;
            /** Format: date-time */
            updated_at: string;
            user_id: string;
            wallet_lock_id?: string;
        };
        ProblemFieldError: {
            code: string;
            message: string;
        };
        ProblemResponse: {
            code: string;
            errors?: components["schemas"]["ProblemFieldError"][];
            message: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}